NetworkManager
Which use Almofire, with generic class and codable protocol.
Created static func fetch, which fetch from urlString. Also has a completion block of escaping type, which return the result of T type and some Network Error and I defined as Void.
Is going to create enum for NetworkError and it has a 2 cases.
I used AF network call and created request where we gonna use the overload which takes a urlString as end point. Then it's returns responseDecode of T.self that returns a response. Than we gonna check if resp.error not equal nil if it's true we gonna say completion send a failure with InvalidResponse, also print it and return.
Check one more thing if let payload, and if there is any data in the response.value, we gonna completion send success that can be decodable & encodable which is payload conforming to codable type and then return.
Alos we gonna have completion for in general cases if payload nil we send failure saying it's another response.
Thats all for Network layer.